feat: Gestiondes pages en fonction du statut loggrdIn + mise en place du logout

+++ delete-article.php
@@ -1,24 +1,26 @@
 <?php
-   
-    /**
-     * @var ArticleDAO
-     */
+    require __DIR__.'/database/database.php';
+    $authDAO = require './database/models/AuthDAO.php';
+    $currentUser = $authDAO->isLoggedIn();
+    var_dump($currentUser);
 
-    $article = require_once './database/models/ArticleDAO';
+    if(!$currentUser) {
+        header('Location: /auth-login.php');
+    } else {
+        /**
+         * @var ArticleDAO 
+         */
+        $articleDAO = require_once './database/models/ArticleDAO.php';
 
+        $articles = [];
 
-    $pdo = require_once "./database/database.php";
-    $statement = $pdo->prepare('DELETE FROM article WHERE id=:id');
+        $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+        $idArticle = $_GET['id'] ?? '';
 
+        
+        if($idArticle) {
+            $articleDAO->deleteOne($idArticle);
+        }
 
-    $filename = __DIR__.'/data/articles.json';
-    $articles = [];
-
-    $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-    $idArticle = $_GET['id'] ?? '';
-
-    if($idArticle) {
-       $articleDAO->deleteOne($idArticle);
+        header('Location: /');
     }
-
-    header('Location: /');
